Defect/de5503 mobile podcast alignment (#173)
* DE5503 - Mobile podcast alignment
Add padding to podcast image on mobile.

* DE5503 - Mobile podcast alignment
Remove CSS sizing constraints for podcast image (not needed when using imgix).

* DE5503 - Mobile podcast alignment
Remove unneccessary podcast text class and duplicated CSS - this effect can be achieved with a utility class.

* DE5503 - Mobile podcast alignment
Adjust layout of podcast artwork.

* DE5503 - Mobile podcast alignment
Clear out extraneous ampersands.

* DE5503 - Mobile podcast alignment
Add styles for jumbotron to account for the podcast image size change.

* DE5503 - Mobile podcast alignment
Remove breakpoint targeting 375px - it's an arbitrary number.

* DE5503 - Mobile podcast alignment
Refactor background image from inline to background image.

* DE5503 - Mobile podcast alignment
Remove extraneous row class - it was messing up the padding since it didn't include any col-* classses.

* DE5503 - Mobile podcast alignment
Nest p tag (with layout class) inside conditional to hide potential empty p tags from DOM if apple podcast link isn't added.

* DE5503 - Mobile podcast alignment
Change layout of podcast art - because we're using a background image and want the content to adjust the container height as needed I had to strip out the flex solution. This allows for that little "cut-out" that appears when content is short and on a desktop sized screen.

* DE5503 - Mobile podcast alignment
Because the container will grow as the content does, setting a min-height is unneccessary. It defaults to the height it needs to be.

* DE5503 - Mobile podcast alignment
Removing unused CSS.

* DE5503 - Mobile podcast alignment
Adjust padding for the content portion of jumbotron so the transparent black background fills it up.

* DE5503 - Mobile podcast alignment
Account for imgix sizes attribute.